+++ src/Application/Rule/Rules/ConditionRule.php
@@ -5,24 +5,17 @@ namespace ArtARTs36\MergeRequestLinter\Application\Rule\Rules;
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionOperator;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDecorator;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Counter;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\NullCounter;
 
-class ConditionRule implements RuleDecorator
+class ConditionRule extends OneRuleDecoratorRule
 {
     public function __construct(
-        private Rule $rule,
+        Rule $rule,
         private ConditionOperator $operator,
         private readonly Counter $skippedRules = new NullCounter(),
     ) {
-        //
-    }
-
-    public function getName(): string
-    {
-        return $this->rule->getName();
+        parent::__construct($rule);
     }
 
     public function lint(MergeRequest $request): array
@@ -34,15 +27,5 @@ class ConditionRule implements RuleDecorator
         }
 
         return $this->rule->lint($request);
-    }
-
-    public function getDefinition(): RuleDefinition
-    {
-        return $this->rule->getDefinition();
-    }
-
-    public function getDecoratedRules(): array
-    {
-        return [$this->rule];
     }
 }