+++ src/Domain/Linter/LintState.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Domain\Linter;
+
+enum LintState: int
+{
+    case Success = 1;
+    case Fail = 2;
+}