+++ src/Presentation/Console/Printers/NotePrinter.php
@@ -6,12 +6,13 @@ use ArtARTs36\MergeRequestLinter\Domain\Note\Note;
 use ArtARTs36\MergeRequestLinter\Domain\Note\NoteSeverity;
 use ArtARTs36\MergeRequestLinter\Presentation\Console\Contracts\TablePrinter;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
+use ArtARTs36\Str\Facade\Str;
 use Symfony\Component\Console\Helper\TableCell;
 use Symfony\Component\Console\Helper\TableCellStyle;
 
 class NotePrinter
 {
-    private const HEADERS = ['#', 'Note'];
+    private const HEADERS = ['#', 'Severity', 'Note'];
 
     /**
      * @param Arrayee<int, Note> $notes
@@ -26,7 +27,12 @@ class NotePrinter
                     'fg' => 'red',
                 ]),
             ],
-            NoteSeverity::Normal->value => [],
+            NoteSeverity::Warning->value => [
+                'style' => new TableCellStyle([
+                    'fg' => 'yellow',
+                ]),
+            ],
+            NoteSeverity::Error->value => [],
         ];
 
         $counter = 0;
@@ -35,9 +41,12 @@ class NotePrinter
         foreach ($notes as $note) {
             ++$counter;
 
+            $cellOptions = $tableCellOptions[$note->getSeverity()->value];
+
             $table[] = [
-                new TableCell("$counter", $tableCellOptions[$note->getSeverity()->value]),
-                new TableCell($note->getDescription(), $tableCellOptions[$note->getSeverity()->value]),
+                new TableCell("$counter", $cellOptions),
+                new TableCell(Str::upFirstSymbol($note->getSeverity()->value), $cellOptions),
+                new TableCell($note->getDescription(), $cellOptions),
             ];
         }