+++ tests/Feature/Console/Command/DumpCommandTest.php
@@ -5,11 +5,10 @@ namespace ArtARTs36\MergeRequestLinter\Tests\Feature\Console\Command;
 use ArtARTs36\MergeRequestLinter\Application\Rule\Dumper\RuleDumper;
 use ArtARTs36\MergeRequestLinter\Application\Rule\TaskHandlers\DumpTaskHandler;
 use ArtARTs36\MergeRequestLinter\Presentation\Console\Command\DumpCommand;
+use ArtARTs36\MergeRequestLinter\Tests\Mocks\CommandTester;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockConfigResolver;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\SuccessRule;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
-use ArtARTs36\Str\Facade\Str;
-use Symfony\Component\Console\Tester\CommandTester;
 
 final class DumpCommandTest extends TestCase
 {
@@ -34,9 +33,6 @@ final class DumpCommandTest extends TestCase
 
         $tester->assertCommandIsSuccessful();
 
-        self::assertStringContainsString(
-            '1 Success rule ArtARTs36\MergeRequestLinter\Tests\Mocks\SuccessRule',
-            Str::deleteUnnecessarySpaces($tester->getDisplay())
-        );
+        $tester->assertDisplayContainsString('success_rule Success rule true');
     }
 }