+++ tests/Unit/Domain/Linter/LintResultTest.php
@@ -3,6 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Domain\Linter;
 
 use ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult;
+use ArtARTs36\MergeRequestLinter\Domain\Linter\LintState;
 use ArtARTs36\MergeRequestLinter\Shared\Time\Duration;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\EmptyNote;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
@@ -24,26 +25,6 @@ final class LintResultTest extends TestCase
     }
 
     /**
-     * @dataProvider providerForTestIsFail
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult::isFail
-     */
-    public function testIsFail(LintResult $result, bool $expected): void
-    {
-        self::assertEquals($expected, $result->isFail());
-    }
-
-    /**
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult::fail
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult::__construct
-     */
-    public function testFail(): void
-    {
-        $result = LintResult::fail(new EmptyNote(), new Duration(0.12));
-
-        self::assertFalse($result->state);
-    }
-
-    /**
      * @covers \ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult::successWithNote
      * @covers \ArtARTs36\MergeRequestLinter\Domain\Linter\LintResult::__construct
      */
@@ -51,6 +32,6 @@ final class LintResultTest extends TestCase
     {
         $result = LintResult::successWithNote(new EmptyNote(), new Duration(0.12));
 
-        self::assertTrue($result->state);
+        self::assertEquals(LintState::Success, $result->state);
     }
 }